# openapi.yaml
openapi: 3.0.3
info:
  title: Alauda API
  version: 1.0.0
  description: |
    🚀 **Alauda API - Multi-Service Integration Platform**
    
    All-in-one API for media downloads, social media content, music identification, payment processing and more. Built for developers who need reliable, fast and affordable integrations.
    
    ## 🎯 Features
    
    **Media Downloads:**
    - 🎵 Spotify - Search and download tracks
    - 📹 YouTube - Video and audio downloader
    - 🎬 TikTok - Download videos without watermark
    - 📸 Instagram - Download posts, reels, stories
    - 📘 Facebook - Video downloader
    
    **Music Services:**
    - 🎤 Shazam - Identify any song
    - 📝 Lyrics - Search song lyrics worldwide
    
    **Utilities:**
    - 🆔 CPF Validator - Brazilian tax ID validation
    - 🖼️ Background Remover - AI-powered image processing
    - 💬 WhatsApp Utils - Formatting and validation tools
    
    **Payment Processing:**
    - 💳 MercadoPago - Brazil and LATAM payments
    - 📱 M-Pesa - Mozambique mobile money (Vodacom)
    - 📱 E-Mola - Mozambique mobile money (Movitel)
    
    ## 🌍 Developed in Mozambique 🇲🇿 | Available Worldwide
    
    **Created by:** Zeus Lykraios 💎
    
    ## 🔑 Authentication
    
    All endpoints (except `/` and `/health`) require an API key sent in the `X-API-Key` header.
    
    Get your free API key at: https://alauda-api.topazioverse.com.br/api/auth/register
    
    ## 💰 Credits System
    
...
    
    ## 📊 Rate Limiting
    
...
    
  contact:
...
servers:
...
security:
  - ApiKeyAuth: []
paths:
...
  /api/youtube/search:
    post:
      tags:
        - Media Downloads
      summary: Search YouTube Videos
      description: Search videos on YouTube by query (free, no external API key needed)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - query
              properties:
                query:
                  type: string
                  example: "Rick Astley"
                max_results:
                  type: integer
                  minimum: 1
                  maximum: 25
                  default: 10
      responses:
        '200':
          description: Search results
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      query:
                        type: string
                      count:
                        type: integer
                      videos:
                        type: array
                        items:
                          type: object
                          properties:
                            id:
                              type: string
                            title:
                              type: string
                            url:
                              type: string
                              format: uri
                            thumbnail:
                              type: string
                              format: uri
                            duration:
                              type: string
                            views:
                              type: string
                            published:
                              type: string
                            channel:
                              type: string
                  credits_remaining:
                    type: integer
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '402':
          $ref: '#/components/responses/InsufficientCredits'
        '500':
          $ref: '#/components/responses/ServerError'
...
  /api/youtube/download:
    post:
      tags:
        - Media Downloads
      summary: Download YouTube Video/Audio
      description: Download videos or audio from YouTube
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - url
              properties:
                url:
                  type: string
                  example: "https://www.youtube.com/watch?v=dQw4w9WgXcQ"
                format:
                  type: string
                  enum: [video, audio]
                  default: video
                quality:
                  type: string
                  enum: [highest, lowest, 720p, 1080p]
                  default: highest
      responses:
        '200':
          description: Download ready
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  download:
                    type: object
                  credits_remaining:
                    type: integer
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: |
        API key for authentication. Get yours at: https://alauda-api.topazioverse.com.br/api/auth/register
  schemas:
...
    Error:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
        message:
          type: string
    ValidationError:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
          example: "Validation error"
        errors:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
              message:
                type: string
  responses:
    Unauthorized:
      description: Invalid or missing API key
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            success: false
            error: "API Key inválida ou expirada"
    ValidationError:
      description: Invalid request parameters
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ValidationError'
    InsufficientCredits:
      description: Not enough credits
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            success: false
            error: "Créditos insuficientes"
            required: 5
            available: 2
    ServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            success: false
            error: "Erro interno do servidor"